# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/room/index": {
            "get": {
                "tags": [
                    "Type Room"
                ],
                "summary": "Read all the types romms",
                "description": "Read all the types romms",
                "operationId": "roomAll",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/room/accommodations/{id}": {
            "get": {
                "tags": [
                    "Type Room"
                ],
                "summary": "Read the type of room",
                "description": "Read the type of room",
                "operationId": "roomGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id type room",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}